+++ packages/core/src/storybook/stand-alone-documentaion/welcome/Abilities/Abilities.tsx
@@ -8,10 +8,10 @@ export const Abilities = () => (
     <Ability title="Getting started" imageSrc={gettingStarted} linkHref="/?path=/docs/getting-started--docs">
       Instructions and welcome to the monday.com OS design system
     </Ability>
-    <Ability title="Foundations" imageSrc={foundations} linkHref="/?path=/docs/foundations-colors--docs">
+    <Ability title="Foundations" imageSrc={foundations} linkHref="/?path=/docs/foundations-colors--colors">
       All information about colors, typography, spacing, and icons
     </Ability>
-    <Ability title="Components" imageSrc={components} linkHref="/?path=/docs/buttons">
+    <Ability title="Components" imageSrc={components} linkHref="/?path=/docs/catalog--docs">
       All the information and guidelines you’ll ever need on each component
     </Ability>
   </div>