+++ packages/icons/src/iconsMetaData.ts
@@ -13,6 +13,13 @@ export default [
   // DO NOT REMOVE THIS FOLLOWING LINE!
   // plop_marker:icon_metadata
   {
+      name: "MinusSmall",
+      file: "MinusSmall.svg",
+      description: "Minus icon for time counter",
+      tags: "Minus, Small, Counter, Remove, Divider, Subtract"
+  },
+
+  {
       name: "ClassicFolder",
       file: "ClassicFolder.svg",
       description: "Typically used when grouping a few items together under one subject",